refactor(sync): post-review fixes (Phase 1)
Addresses self+independent code review findings on 0556d61.

Correctness:
- fetchAndApplyDeltaChanges: applier + KEY_LAST_SYNC_TIME upsert now in one
  transaction. Previously a crash between the two left applied rows with
  stale lastSync — next run would re-pull and (once Phase 5 lands) the
  updated_at guard would silently skip newer server changes.
- SyncManager.runOnce(Manual) now awaits the mutex instead of returning a
  fake success — manual triggers from the UI no longer get misleading
  outcomes when a periodic run is in flight.
- Rejected-not-in-batch ids no longer reset the retry trajectory: the
  `?: 0L` fallback was silently treating server contract violations as a
  fresh first attempt. Now logged via Napier.e and skipped.
- Throwable classifier handles UnresolvedAddressException (DNS) → NETWORK.

Architecture:
- pendingCount Flow and last-sync meta writes (success/failure ts + reason)
  moved from SyncManager to SyncRepository. Cuts SyncManager constructor
  deps 7 → 3 (syncRepository, actionLogger, dispatchers). Storage access
  no longer leaks into the manager.
- TransactionRunner.transactional generic dropped — single caller shape
  was Unit-only; removed the SENTINEL/UNCHECKED_CAST dance.
- LocalSyncStatus.FAILED removed entirely (no callers — was never the right
  semantic; retry counters cover the case).

Cleanness:
- scheduleBatchRetry / scheduleSingleRetry collapsed into a single
  scheduleRetry(entity, id, currentAttempt, now, reason) helper over a
  private RetryEntity enum. Removes ~70 lines of duplicated per-entity
  retry plumbing. Attempt lookup at reject time goes through a
  buildAttemptsLookup map keyed by (entity, id).
- SyncStatus.Success counters aligned to Long (matches Flow<Long>
  pendingCount and SQLDelight count()).

+++ shared/sync-manager/src/commonMain/kotlin/ru/mirea/toir/sync/domain/SyncStatus.kt
@@ -9,8 +9,8 @@ sealed interface SyncStatus {
     data object Running : SyncStatus
     data class Success(
         val finishedAt: Instant,
-        val pushedCount: Int,
-        val uploadedPhotoCount: Int,
+        val pushedCount: Long,
+        val uploadedPhotoCount: Long,
     ) : SyncStatus
     data class Failed(
         val finishedAt: Instant,
